refactor(*): code moved outside init file
